+++ library/src/main/java/de/mrapp/android/tabswitcher/layout/tablet/TabletTabRecyclerAdapter.java
@@ -29,14 +29,12 @@ import android.view.ViewGroup;
 
 import de.mrapp.android.tabswitcher.Layout;
 import de.mrapp.android.tabswitcher.R;
-import de.mrapp.android.tabswitcher.Tab;
 import de.mrapp.android.tabswitcher.TabSwitcher;
 import de.mrapp.android.tabswitcher.iterator.ItemIterator;
 import de.mrapp.android.tabswitcher.layout.AbstractTabRecyclerAdapter;
 import de.mrapp.android.tabswitcher.layout.AbstractTabViewHolder;
 import de.mrapp.android.tabswitcher.model.AbstractItem;
 import de.mrapp.android.tabswitcher.model.AddTabItem;
-import de.mrapp.android.tabswitcher.model.Model;
 import de.mrapp.android.tabswitcher.model.TabItem;
 import de.mrapp.android.tabswitcher.model.TabSwitcherModel;
 import de.mrapp.android.tabswitcher.model.TabSwitcherStyle;
@@ -49,8 +47,7 @@ import de.mrapp.android.util.ViewUtil;
  * @author Michael Rapp
  * @since 1.0.0
  */
-public class TabletTabRecyclerAdapter extends AbstractTabRecyclerAdapter
-        implements Tab.Callback, Model.Listener {
+public class TabletTabRecyclerAdapter extends AbstractTabRecyclerAdapter {
 
     /**
      * The view type of a button, which allows to add a new tab.